+++ service/src/test/java/org/sunbird/service/user/impl/UserExternalIdentityServiceImplTest.java
@@ -7,6 +7,8 @@ import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
+
+import org.codehaus.jackson.map.ObjectMapper;
 import org.junit.Assert;
 import org.junit.Before;
 import org.junit.Test;
@@ -20,6 +22,7 @@ import org.sunbird.cassandra.CassandraOperation;
 import org.sunbird.cassandraimpl.CassandraOperationImpl;
 import org.sunbird.helper.ServiceFactory;
 import org.sunbird.keys.JsonKey;
+import org.sunbird.model.user.UserDeclareEntity;
 import org.sunbird.request.RequestContext;
 import org.sunbird.response.Response;
 import org.sunbird.service.user.UserExternalIdentityService;
@@ -37,7 +40,7 @@ import org.sunbird.service.user.UserExternalIdentityService;
   "javax.crypto.*"
 })
 public class UserExternalIdentityServiceImplTest {
-
+  ObjectMapper mapper = new ObjectMapper();
   private static CassandraOperation cassandraOperationImpl = null;
 
   @Before
@@ -84,5 +87,54 @@ public class UserExternalIdentityServiceImplTest {
     List<Map<String, String>> externalIds =
         userExternalIdentityService.getExternalIds("userId", true, new RequestContext());
     Assert.assertNotNull(externalIds);
+  }
+
+
+  @Test
+  public void testConvertSelfDeclareFieldsToExternalIds() {
+    Map<String, Object> selfDeclaredFields = getSelfDeclareFields();
+    List<Map<String, String>> externalIds =
+            UserExternalIdentityServiceImpl.convertSelfDeclareFieldsToExternalIds(selfDeclaredFields);
+    String declaredEmail = "";
+    String declaredPhone = "";
+    for (Map<String, String> extIdMap : externalIds) {
+      if (JsonKey.DECLARED_EMAIL.equals((String) extIdMap.get(JsonKey.ORIGINAL_ID_TYPE))) {
+        declaredEmail = (String) extIdMap.get(JsonKey.ORIGINAL_EXTERNAL_ID);
+      }
+      if (JsonKey.DECLARED_PHONE.equals((String) extIdMap.get(JsonKey.ORIGINAL_ID_TYPE))) {
+        declaredPhone = (String) extIdMap.get(JsonKey.ORIGINAL_EXTERNAL_ID);
+      }
+    }
+
+    Assert.assertEquals("abc@tenant.com", declaredEmail);
+    Assert.assertEquals("999999999", declaredPhone);
+  }
+
+  @Test
+  public void testConvertExternalFieldsToSelfDeclareFields() {
+    Map<String, Object> declaredFeilds = getSelfDeclareFields();
+    List<Map<String, String>> externalIds =
+            UserExternalIdentityServiceImpl.convertSelfDeclareFieldsToExternalIds(declaredFeilds);
+    Map<String, Object> resultDeclaredFields =
+            UserExternalIdentityServiceImpl.convertExternalFieldsToSelfDeclareFields(externalIds);
+    Assert.assertEquals(
+            declaredFeilds.get(JsonKey.USER_ID), resultDeclaredFields.get(JsonKey.USER_ID));
+    Assert.assertEquals(
+            ((Map<String, Object>) declaredFeilds.get(JsonKey.USER_INFO)).get(JsonKey.DECLARED_EMAIL),
+            ((Map<String, Object>) resultDeclaredFields.get(JsonKey.USER_INFO))
+                    .get(JsonKey.DECLARED_EMAIL));
+  }
+
+  private Map<String, Object> getSelfDeclareFields() {
+    UserDeclareEntity userDeclareEntity = new UserDeclareEntity();
+    userDeclareEntity.setUserId("userid");
+    userDeclareEntity.setOrgId("org");
+    userDeclareEntity.setPersona(JsonKey.TEACHER_PERSONA);
+    Map<String, Object> userInfo = new HashMap<>();
+    userInfo.put(JsonKey.DECLARED_EMAIL, "abc@tenant.com");
+    userInfo.put(JsonKey.DECLARED_PHONE, "999999999");
+    userDeclareEntity.setUserInfo(userInfo);
+    Map<String, Object> selfDeclaredMap = mapper.convertValue(userDeclareEntity, Map.class);
+    return selfDeclaredMap;
   }
 }
